Fix: CategoryList에 css variable 적용
