+++ src/components/CategoryHeader/CategoryList/index.tsx
@@ -1,4 +1,4 @@
-import * as S from './CategoryList.style'
+import * as S from '../CategoryHeader.style'
 
 export interface CategoryListProps {
   selectedCategory: string
@@ -11,7 +11,7 @@ const CategoryList = ({ selectedCategory, categoryList }: CategoryListProps) => 
   const { All: AllCategoryTag, ...restCategoryList } = categoryList
   return (
     <S.CategoryList>
-      <S.AllCategoryTag to="/?category=All" active>
+      <S.AllCategoryTag to="/?category=All" active={'All' === selectedCategory}>
         전체 {AllCategoryTag}개의 포스팅
       </S.AllCategoryTag>
       {Object.entries(restCategoryList).map(([name, count]) => (
